TEST CASE - 954.5 RPM
Right now the currentStep variable is able to increment
TOO FAST. It seems that the processor was so busy processing
timer interrupts (even though the ISR was empty), that
it has no time to run the main loop.

First, I need to point out that the main loop is a little retarded
right now. My intention in the previous release was to have the
loop only increment after a PWM cycle has completed, but I never
fully implemented it. The intention of the loop was to have it loop
a certain number of PWM cycles, and then change the PWM value just
as it begins the next cycle. This is not happening. It is freely changing
the value.

Here are some considerations:

1) Beyond about 250 RPM motor shakes and does not turn, however
   we are able to flick the motor and have it run up to about 950
   RPM.  In the future we can ramp up the motor speed to prevent this
   which is something which could either be done in our ESC firmware,
   or by the user.
2) Beyond about 1000 RPM I am unable to sucessfully get the motor to
   move, even with flicking. It is not clear if higher speeds can be
   acheived if it is ramped up to speed. If there is more inertia
   created by attaching somethign to the motor spindle, and the motor
   is ramped, it may be possible to acheive higher speeds.

3) There is always the risk that if the motor is turning at higher speed
   and and some force stops or resists the spindle, that the motor will
   not be able to recover. Without RPM feedback, we would have no way
   to know that this occured, outside of detecting something wrong
   at the system level.

WHAT CHANGED IN THE CODE

1) I disabled the timer Interrupt
    I thought I did this before, but ...
    This made for a tremendous improvement.
2) Retardedly implemented some delay in incrementing the
   currentStep variabled.