Created infrastructure to allow update of ISR structure later if busy.
ISR changover to the next table only occurs at the and of the PWM cycle.
From the time that the alternate table has been initialized to the time
that the ISR switched to that table, no further changes can be made.

To address this issue, changes were made which allow the update to be post
poned until after the ISR has brought over the previous changes.

1) Created updateISR which does the updating of the ISR data structure.
2) Made the update method only call updateISR only if the an update is
   required and the ISR is free to accept more changes.
3) Created a 'busy' method for determining if the ISR can not accept
   more changes into its structure.
4) Created -updateOutstanding property which indicates if an update is
   waiting to occur (i.e. we are waiting for the ISR to be free to accept
   a change.