ISR Routine now works and spins motor at constant slow speed.
This was debugging of the ISR. I still need to make the motor move smooth
and give it variable speed. whew !

There was a lot of debug and experimental code which needs to be pulled out
or properly organized.  I checked this in rough because this is a critical
release and I dont want to risk messing it up.