bldcPwm = Solved motor jitter during slow turn
Problem was that as it turned very slowely, the motor
would shake breifly in certain positions. This
was cause by the PWM going to double frequency
at certain pwm indexes (i.e. 10 and 11). At these
indexes the isrExit_WAIT was active,  but after teh
timer expired we did not clear the interrupt, making
the next full ISR exit, re-enter the isr immediatley
this skipping the delay. The skipped delay caused the
higher frequency.